Transposition Table: use uint16 for score

+++ move_generation_test.go
@@ -217,7 +217,7 @@ func TestGenerateQuiescentMoves(t *testing.T) {
 	testBoard.SetPieceAtSquare(SQUARE_A8, BLACK_MASK|QUEEN_MASK)
 
 	moves := make([]Move, 64)
-	moveScores := make([]int, len(moves))
+	moveScores := make([]int16, len(moves))
 	end := GenerateQuiescentMoves(&testBoard, moves[:], moveScores[:], 0)
 	assert.Equal(t, 3, end)
